backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "Annotation-Backend"
version = "1.0"
description = ""
requires-python = ">=3.11.0,<3.12"
dependencies = [
    "torch==2.5.1",
    "torchvision==0.20.1",
    "numpy==1.24.4",
    "opencv-python>=4.7.0",
    "tqdm==4.66.5",
    "hydra-core==1.3.2",
    "iopath==0.1.10",
    "pillow==10.0.0",
]

[project.optional-dependencies]
interactive-UI = [
    "fastapi==0.119.1",
    "uvicorn==0.38.0",
    "pydantic==2.12.3",
    "python-multipart==0.0.20",
    "av==13.1.0",
    "dataclasses-json==0.6.7",
    "eva-decord==0.6.1",
    "imagesize==1.4.1",
    "pycocotools==2.0.8",
    "strawberry-graphql[fastapi]==0.243.0",
]
dev = [
    "black==24.2.0",
    "usort==1.0.2",
    "ufmt==2.0.0b2",
    "fvcore==0.1.5.post20221221",
    "pandas==2.2.2",
    "scikit-image==0.24.0",
    "tensorboard>=2.17.0",
    "pycocotools>=2.0.8",
    "tensordict>=0.6.0",
    "submitit>=1.5.1",
]

[tool.setuptools.packages.find]
include = ["*"]